[FEAT] 거래 정보를 별도의 테이블로 분리

+++ src/main/java/com/onebucket/domain/boardManage/dto/internal/post/CreateMarketPostDto.java
@@ -28,7 +28,7 @@ public class CreateMarketPostDto extends CreatePostDto {
     private Long price;
     private Long count;
 
-    private int wanted;
+    private Long wanted;
     private String location;
     private LocalDateTime dueDate;
 }
