+++ src/main/java/com/onebucket/domain/boardManage/dto/request/RequestCreateMarketPostDto.java
@@ -21,6 +21,6 @@ import lombok.Setter;
 @Setter
 public class RequestCreateMarketPostDto extends RequestCreatePostDto {
     private String item;
-    private int wanted;
+    private Long wanted;
     private String location;
 }
