+++ src/main/java/com/onebucket/domain/boardManage/dto/response/ResponseMarketPostDto.java
@@ -23,6 +23,6 @@ import lombok.experimental.SuperBuilder;
 @Setter
 public class ResponseMarketPostDto extends ResponsePostDto {
     private String item;
-    private int wanted;
+    private Long wanted;
     private String location;
 }
